+++ NEWS.md
@@ -12,6 +12,7 @@ Development
   - New visualizations backup [#14745](https://github.com/CartoDB/cartodb/pull/14745)
   - Restore visualizations from new backup [#14764](https://github.com/CartoDB/cartodb/pull/14764)
 - Tag search [#14777](https://github.com/CartoDB/cartodb/pull/14777)
+- Ghost tables event trigger creation [#14697](https://github.com/CartoDB/cartodb/issues/14697)
 
 ### Bug fixes / enhancements
 - Add link to Help Center to invitation emails